✨ :: login view setup
