+++ DMSport/Sources/Scene/Login/Ui/Sign/MomViewConfigure.swift
@@ -1,0 +1,11 @@
+//
+//  MomViewConfigure.swift
+//  DMSport-iOS
+//
+//  Created by 박준하 on 2022/10/18.
+//  Copyright © 2022 com.DMS. All rights reserved.
+//
+
+import Foundation
+
+extension MomView {
